+++ pyannote/database/__init__.py
@@ -3,7 +3,7 @@
 
 # The MIT License (MIT)
 
-# Copyright (c) 2016 CNRS
+# Copyright (c) 2016-2020 CNRS
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -27,21 +27,26 @@
 # Hervé BREDIN - http://herve.niderb.fr
 
 
-"""
-
-"""
+"""pyannote.database"""
 
 import sys
 from pkg_resources import iter_entry_points
 
+from typing import Optional, Dict, Set, Text
+
 from .database import Database
 from .database import PyannoteDatabaseException
 
+from .protocol.protocol import Protocol
+from .protocol.protocol import ProtocolFile
+from .protocol.protocol import Subset
+from .protocol.protocol import Preprocessors
+
 DATABASES = dict()
-TASKS = dict()
+TASKS: Dict[Text, Set[Text]] = dict()
 
 # load databases from entry points
-for o in iter_entry_points(group='pyannote.database.databases', name=None):
+for o in iter_entry_points(group="pyannote.database.databases", name=None):
 
     database_name = o.name
 
@@ -59,6 +64,7 @@ for o in iter_entry_points(group='pyannote.database.databases', name=None):
 
 # parse pyannote.database configuration file, looking for custom protocols
 from .custom import add_custom_protocols
+
 DATABASES, TASKS = add_custom_protocols()
 
 
@@ -84,18 +90,13 @@ def get_databases(task=None):
     return sorted(TASKS.get(task, []))
 
 
-def get_database(database_name, preprocessors={}, **kwargs):
+def get_database(database_name, **kwargs):
     """Get database by name
 
     Parameters
     ----------
     name : str
         Database name.
-    preprocessors : dict or (key, preprocessor) iterable
-        When provided, each protocol item (dictionary) are preprocessed, such
-        that item[key] = preprocessor(item). In case 'preprocessor' is not
-        callable, it should be a string containing placeholder for item keys
-        (e.g. {'audio': '/path/to/{uri}.wav'})
 
     Returns
     -------
@@ -108,25 +109,25 @@ def get_database(database_name, preprocessors={}, **kwargs):
 
     except KeyError as e:
 
-        if database_name == 'X':
+        if database_name == "X":
             msg = (
-                'Could not find any meta-protocol. Please refer to '
-                'pyannote.database documentation to learn how to define them: '
-                'https://github.com/pyannote/pyannote-database'
+                "Could not find any meta-protocol. Please refer to "
+                "pyannote.database documentation to learn how to define them: "
+                "https://github.com/pyannote/pyannote-database"
             )
         else:
             msg = (
                 'Could not find any protocol for "{name}" database. Please '
-                'refer to pyannote.database documentation to learn how to '
-                'define them: https://github.com/pyannote/pyannote-database'
+                "refer to pyannote.database documentation to learn how to "
+                "define them: https://github.com/pyannote/pyannote-database"
             )
             msg = msg.format(name=database_name)
         raise ValueError(msg)
 
-    return database(preprocessors=preprocessors, **kwargs)
+    return database(**kwargs)
 
 
-def get_protocol(name, preprocessors={}, progress=False, **kwargs):
+def get_protocol(name, preprocessors: Optional[Preprocessors] = None) -> Protocol:
     """Get protocol by full name
 
     name : str
@@ -136,8 +137,6 @@ def get_protocol(name, preprocessors={}, progress=False, **kwargs):
         that item[key] = preprocessor(item). In case 'preprocessor' is not
         callable, it should be a string containing placeholder for item keys
         (e.g. {'audio': '/path/to/{uri}.wav'})
-    progress : bool, optional
-        Defaults to False.
 
     Returns
     -------
@@ -145,12 +144,11 @@ def get_protocol(name, preprocessors={}, progress=False, **kwargs):
         Protocol instance
     """
 
-    database_name, task_name, protocol_name = name.split('.')
-    database = get_database(database_name,
-                            preprocessors=preprocessors,
-                            **kwargs)
-    protocol = database.get_protocol(task_name, protocol_name,
-                                     progress=progress)
+    database_name, task_name, protocol_name = name.split(".")
+    database = get_database(database_name)
+    protocol = database.get_protocol(
+        task_name, protocol_name, preprocessors=preprocessors
+    )
     return protocol
 
 
@@ -164,6 +162,8 @@ from .util import get_annotated
 from .util import get_unique_identifier
 from .util import get_label_identifier
 
+
 from ._version import get_versions
-__version__ = get_versions()['version']
+
+__version__ = get_versions()["version"]
 del get_versions