+++ pyannote/database/protocol/__init__.py
@@ -3,7 +3,7 @@
 
 # The MIT License (MIT)
 
-# Copyright (c) 2016-2018 CNRS
+# Copyright (c) 2016-2020 CNRS
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
 
-
+from .protocol import Protocol
 from .collection import CollectionProtocol
 from .speaker_diarization import SpeakerDiarizationProtocol
 from .speaker_spotting import SpeakerSpottingProtocol